requirements.txt
Pillow
impacket
netifaces
